chore: add maintenance notice (#64)
<!--
Thank you for contributing to the project! 💜
Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

We are no longer using this Github Action and will not be maintaining it
moving forward. This updates the README to say that.